services/equipment/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
firebase-admin>=6.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
qrcode>=7.4.0
Pillow>=10.0.0
nanoid>=2.0.0
redis>=5.0.0
